refactor: create universal throttle manager
