+++ app/src/main/java/exh/util/ThrottleManager.kt
@@ -1,16 +1,18 @@
-package exh.eh
+package exh.util
 
 import kotlinx.coroutines.delay
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.seconds
 
-class EHentaiThrottleManager(
+class ThrottleManager(
     private val max: Duration = THROTTLE_MAX,
     private val inc: Duration = THROTTLE_INC,
+    private val initial: Duration = Duration.ZERO
 ) {
     private var lastThrottleTime = Duration.ZERO
-    var throttleTime = Duration.ZERO
+
+    var throttleTime = initial
         private set
 
     suspend fun throttle() {
@@ -30,7 +32,7 @@ class EHentaiThrottleManager(
 
     fun resetThrottle() {
         lastThrottleTime = Duration.ZERO
-        throttleTime = Duration.ZERO
+        throttleTime = initial
     }
 
     companion object {
